requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
exif==1.6.0
idna==3.4
piexif==1.1.3
Pillow==9.5.0
plum-py==0.8.6
requests==2.31.0
urllib3==2.0.2
